% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gene.R
\name{get_gene}
\alias{get_gene}
\title{Get Gene}
\usage{
get_gene(accession)
}
\arguments{
\item{accession}{a character string of a UniProt accession identification code}
}
\value{
a character string of the gene ID
}
\description{
Get Gene
}
\examples{
{
gene <- get_gene('Q28021')

}
}
