% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_keywords.R
\name{get_keywords}
\alias{get_keywords}
\title{Get Keywords}
\usage{
get_keywords(accession)
}
\arguments{
\item{accession}{a character string of a UniProt accession identification code}
}
\value{
a character vector of keywords
}
\description{
Get Keywords
}
\examples{
{
keywords <- get_keywords('Q28021')

}
}
