% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sequence.R
\name{get_sequence}
\alias{get_sequence}
\title{Get Sequence}
\usage{
get_sequence(accession)
}
\arguments{
\item{accession}{a character string of a UniProt accession identification code}
}
\value{
a character vector of protein sequence
}
\description{
Get Sequence
}
\examples{
{
sequence <- get_sequence('Q28021')

}
}
