% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go_query.R
\name{go_query}
\alias{go_query}
\title{Get Gene Ontology}
\usage{
go_query(accession)
}
\arguments{
\item{accession}{a character string of a UniProt accession identification code}
}
\value{
a tibble of the gene ontology (GO) terms
}
\description{
Get Gene Ontology
}
\examples{
{
GO_res <- go_query('Q28021')

}
}
